Canvas image vs layer

+++ src/Common.Media/Drawing/Constants/ImageLayerDefaults.cs
@@ -11,7 +11,7 @@ public static class ImageLayerDefaults
     public static int Margin { get; set; } = 0;
     public static ImagePosition Position { get; set; } = ImagePosition.Absolute;
     public static ImageEdgeOffset Offset { get; set; } = new(0, 0);
-    public static float Rotation { get; set; } = 0;
+    public static int Rotation { get; set; } = 0;
     public static float Opacity { get; set; } = 1;
     public static int Dpi { get; set; } = DimensionsUtility.DPI.DEFAULT;
 }
